+++ src/frameworks/uikit/ui_device.rs
@@ -170,7 +170,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     if matches!(
         env.bundle.bundle_identifier(),
         "com.apprisetec9.minionjump" | "com.risinghighapps.kingdomprincepro"
-    ) || std::env::var_os("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY").is_some() {
+    ) || crate::env_flag_cached!("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY") {
         return ns_string::get_static_str(env, "iPad");
     }
     // ULTRAHLE_MINIONJUMP_MODEL_END
@@ -193,7 +193,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     if matches!(
         env.bundle.bundle_identifier(),
         "com.apprisetec9.minionjump" | "com.risinghighapps.kingdomprincepro"
-    ) || std::env::var_os("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY").is_some() {
+    ) || crate::env_flag_cached!("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY") {
         return ns_string::get_static_str(env, "iPad");
     }
     // ULTRAHLE_MINIONJUMP_NAME_END
@@ -275,7 +275,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     if matches!(
         env.bundle.bundle_identifier(),
         "com.apprisetec9.minionjump" | "com.risinghighapps.kingdomprincepro"
-    ) || std::env::var_os("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY").is_some() {
+    ) || crate::env_flag_cached!("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY") {
         return UIUserInterfaceIdiomPad;
     }
     // ULTRAHLE_MINIONJUMP_IDIOM_END
@@ -343,7 +343,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     if matches!(
         env.bundle.bundle_identifier(),
         "com.apprisetec9.minionjump" | "com.risinghighapps.kingdomprincepro"
-    ) || std::env::var_os("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY").is_some() {
+    ) || crate::env_flag_cached!("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY") {
         return ns_string::get_static_str(env, "iPad2,1");
     }
     // ULTRAHLE_MINIONJUMP_PLATFORM_END
@@ -361,7 +361,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     if matches!(
         env.bundle.bundle_identifier(),
         "com.apprisetec9.minionjump" | "com.risinghighapps.kingdomprincepro"
-    ) || std::env::var_os("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY").is_some() {
+    ) || crate::env_flag_cached!("TOUCHHLE_FORCE_IPAD_DEVICE_IDENTITY") {
         return ns_string::get_static_str(env, "iPad2,1");
     }
     // ULTRAHLE_MINIONJUMP_HWMODEL_END